채팅방 type에 의한 구분을 해줘야할 것 같음

+++ jae_project/src/main/java/com/springboot/project/data/dto/ProductDTO.java
@@ -27,6 +27,6 @@ public class ProductDTO {
 	private String user_email;
 	private LocalDateTime createdAt;
 	private LocalDateTime updatedAt;
-	private int chatRoomId;
+	private Long chatRoomId;
 	
 }
